Fixed allow anonymous clashes with authorize
https://github.com/SoftUni-Internal/exam-systems-issues/issues/1559
